reduced color count to four, fixed miscellaneous issues
